+++ spec/api_spec.rb
@@ -350,6 +350,64 @@ describe 'openstack-image::api' do
             expect(chef_run).to render_file(file.name).with_content(/^flavor = keystone\+cachemanagement$/)
           end
         end
+
+        context 'keystone authtoken attributes with default values' do
+          it 'sets memcached server(s)' do
+            expect(chef_run).not_to render_file(file.name).with_content(/^memcached_servers = $/)
+          end
+
+          it 'sets memcache security strategy' do
+            expect(chef_run).not_to render_file(file.name).with_content(/^memcache_security_strategy = $/)
+          end
+
+          it 'sets memcache secret key' do
+            expect(chef_run).not_to render_file(file.name).with_content(/^memcache_secret_key = $/)
+          end
+
+          it 'sets cafile' do
+            expect(chef_run).not_to render_file(file.name).with_content(/^cafile = $/)
+          end
+
+          it 'sets insecure' do
+            expect(chef_run).to render_file(file.name).with_content(/^insecure = false$/)
+          end
+
+          it 'sets token hash algorithms' do
+            expect(chef_run).to render_file(file.name).with_content(/^hash_algorithms = md5$/)
+          end
+        end
+
+        context 'keystone authtoken attributes with new values' do
+          it 'sets memcached server(s)' do
+            node.set['openstack']['image']['api']['auth']['memcached_servers'] = 'localhost:11211'
+            expect(chef_run).to render_file(file.name).with_content(/^memcached_servers = localhost:11211$/)
+          end
+
+          it 'sets memcache security strategy' do
+            node.set['openstack']['image']['api']['auth']['memcache_security_strategy'] = 'MAC'
+            expect(chef_run).to render_file(file.name).with_content(/^memcache_security_strategy = MAC$/)
+          end
+
+          it 'sets memcache secret key' do
+            node.set['openstack']['image']['api']['auth']['memcache_secret_key'] = '0123456789ABCDEF'
+            expect(chef_run).to render_file(file.name).with_content(/^memcache_secret_key = 0123456789ABCDEF$/)
+          end
+
+          it 'sets cafile' do
+            node.set['openstack']['image']['api']['auth']['cafile'] = 'dir/to/path'
+            expect(chef_run).to render_file(file.name).with_content(%r{^cafile = dir/to/path$})
+          end
+
+          it 'sets insecure' do
+            node.set['openstack']['image']['api']['auth']['insecure'] = true
+            expect(chef_run).to render_file(file.name).with_content(/^insecure = true$/)
+          end
+
+          it 'sets token hash algorithms' do
+            node.set['openstack']['image']['api']['auth']['hash_algorithms'] = 'sha2'
+            expect(chef_run).to render_file(file.name).with_content(/^hash_algorithms = sha2$/)
+          end
+        end
       end
 
       it 'notifies glance-api restart' do